--- include/ui_components/size.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-struct UISize {
-  float width = 0.0f;
-  float height = 0.0f;
-};
