+++ include/ui_components/ui_component.hpp
@@ -9,7 +9,7 @@
 #include "rect.hpp"
 #include "size.hpp"
 #include "tap_event.hpp"
-#include "ui_key.hpp"
+#include "key.hpp"
 
 using TapListener = std::function<void(const UITapEvent &event)>;
